base_api/requirements.txt
asgiref==3.5.2
boto3==1.24.81
botocore==1.27.81
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Django==3.2
django-cors-headers==3.13.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
faiss-cpu==1.7.2
filelock==3.8.0
huggingface-hub==0.9.1
idna==3.4
jmespath==1.0.1
joblib==1.2.0
numpy==1.21.0
openpyxl==3.0.10
packaging==21.3
pandas==1.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
regex==2022.9.13
requests==2.28.1
s3transfer==0.6.0
sacremoses==0.0.53
sentencepiece==0.1.97
six==1.16.0
sqlparse==0.4.3
tokenizers==0.12.1
torch==1.8.0
tqdm==4.64.1
transformers==4.22.1
typing_extensions==4.3.0
urllib3==1.26.12
